fixed bug (bot printing 2 cats when only 1 needed)
The command invoking text thing (idk what to call this) is also being replaced with ᓚᘏᗢ.
As we have seen `!catify cat` gives out `ᓚᘏᗢ ᓚᘏᗢ` since the `cat` in the `!catify` also gets replaced.
Fix? Before replacing the string, make sure that the string is just the part after the command invoking thing.